Title: 실수 받아 그대로 출력, Time: 92ms, Memory: 9MB, Status: Passed - Codetree

+++ 250411/실수 받아 그대로 출력/enter-real-value-and-print.java
@@ -8,7 +8,7 @@ public class Main {
         BufferedWriter bw = new BufferedWriter(new OutputStreamWriter(System.out));
         StringTokenizer st;
 
-        double a = Integer.parseInt(br.readLine());
+        double a = Double.parseDouble(br.readLine());
         System.out.printf("%.2f", a);
 
         bw.flush();
